backend/requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
boto3==1.20.11
botocore==1.23.11
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.0
decorator==5.1.0
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-rest-knox==4.1.0
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dynaconf==3.1.7
Faker==9.8.0
h11==0.12.0
hvac==0.11.2
idna==3.3
Jinja2==3.0.3
jmespath==0.10.0
loguru==0.5.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.22.3
outcome==1.1.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.2.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
selenium==4.1.0
shortuuid==1.0.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
validators==0.18.2
webdriver-manager==3.5.2
win32-setctime==1.0.3
wsproto==1.0.0
